Isolate Dexcom TLS regression and exercise session expiry
